# Commander: fix parachute trigger