@dagar 2025-02-18
I think there are a few things to untangle here.
 - these actuator_armed flags are not becoming any clearer over time. Now `actuator_armed.lockdown` is simply synonymous with HIL?
 - the usage of NAVIGATION_STATE_TERMINATION

Should we talk this through on the dev call?