@bresch 2025-02-18
> Now actuator_armed.lockdown is simply synonymous with HIL?

To me, `lockdown` should be a safe state where nothing mover. So don't run the motors, don't move servos, don't trigger a parachute. Currently this is required when using HIL, yes. It's apparently also used when in throw mode. 

> the usage of NAVIGATION_STATE_TERMINATION

There, the drone in a special mode where the autopilot decided that the drone can't fly anymore and it's time for last resort measures (parachute, protect payload, self-destruct, ...). We could assume that it's not safe to be next to the drone when this occurs, so it shouldn't be able to override `lockdown`.

Yes, let's discuss that in the DevCall.