@bresch 2025-03-03
During the dev call we said that we should change the output failsafe behavior. Instead of having a failsafe PWM position for each actuator, we could have something more explicit (e.g.: rotors->stop, parachute->trigger, gimbal->look up/close bay). 

However, I can't work on this right now and we need a quick solution because right now the parachute doesn't work on main.